requirements.txt
django==1.4.1
django-grappelli
sorl-thumbnail
south
django-extensions
